backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
email-validator==2.1.0

# Google Gemini
google-generativeai==0.8.3

# Qdrant Vector DB
qdrant-client==1.12.0

# Async support
httpx==0.27.0
aiofiles==24.1.0

# Data processing & scraping
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
playwright==1.48.0

# PDF parsing
PyMuPDF==1.24.13
pypdf==4.3.1
pdf2image==1.17.0

# Document parsing (extended)
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0
openpyxl==3.1.5

# Document export
reportlab==4.2.5
jinja2==3.1.4

# Stripe payments
stripe==11.2.0

# Email service
resend>=2.0.0
# Task Queue & Scheduling
celery==5.4.0
celery[redis]==5.4.0
flower==2.0.1  # Celery monitoring UI

# Retry Logic
tenacity==9.0.0

# Firebase Auth
firebase-admin==6.5.0

# Redis Cache
redis==5.0.1
hiredis==2.3.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0  # for testing API

# Monitoring
sentry-sdk[fastapi]==2.14.0

# Code quality
black==24.8.0
flake8==7.1.1
mypy==1.11.2
